Charge the budget during the walk, parse before casting, round derived edges
All three findings are last round's fixes stopping one step short, in the same
way each time: I corrected the input and not the value derived from it, the
range and not the parse, the loop and not the work inside it.

**The budget was charged around the page, not during it.** `extractPage()`
builds the whole run array before returning, so ticking between pages — or even
per produced run, which was my first attempt — lets a single page holding
millions of text-showing operators spend the entire budget before anything
looks. The unit charged has to be the unit a pathological stream multiplies, so
the tick is now in the content-stream operation loop inside `walk()`, threaded
through the form-XObject recursion with it. The test counts consultations
through an injected clock on a *single-page* fixture, which is the case that
distinguishes charging during the work from charging around it.

**The configured tolerance was cast before it was parsed.** `(float) "one"` is
`0.0`, and zero is a legal tolerance — so a typo silently switched every
cross-check to demanding an exact coordinate match. That is a change in what
the product asserts rather than a configuration error, and nobody would see it
until a check that should have passed did not. The config entry is left uncast
and the provider parses it; the range check the resolver already had is
unchanged. Parsing and range are independent, so the test covers both, and a
word and an empty string are the cases the previous fix let through.

**The derived edges were still raw sums.** Canonicalising the components is not
enough when the check adds them: `601.998 + 10.003` is `612.00100000000009`
while `612 + 0.001` is `612.00099999999998`, so a rectangle sitting exactly on
the permitted boundary was refused for a difference that exists only in the
representation. The right and bottom edges are rounded before comparison, as
the field-schema validator already does. The test asserts the precondition —
that the raw sum really does exceed the boundary and the canonical one does not
— so it cannot quietly stop testing what it was written for.

+++ config/esign.php
@@ -645,8 +645,12 @@ return [
          * tolerance turns the check into a formality that passes on a page whose text has moved.
          * A document may name its own `anchor.tolerance`, which wins over this.
          * See docs/preparation/anchors.md.
+         *
+         * Left uncast on purpose. `(float) "one"` is `0.0`, which is a legal tolerance, so a
+         * casting typo would silently switch every cross-check to demanding an exact match rather
+         * than failing as the configuration mistake it is. The provider parses it.
          */
-        'anchor_cross_check_tolerance' => (float) env('ESIGN_ANCHOR_CROSS_CHECK_TOLERANCE', 1.0),
+        'anchor_cross_check_tolerance' => env('ESIGN_ANCHOR_CROSS_CHECK_TOLERANCE', 1.0),
     ],
 
     /*
